+++ core/acp-ftp/src/main/java/pers/acp/ftp/InitSFtpServer.java
@@ -3,7 +3,9 @@ package pers.acp.ftp;
 import pers.acp.core.CommonTools;
 import pers.acp.core.interfaces.IDaemonService;
 import pers.acp.core.log.LogFactory;
+import pers.acp.ftp.base.InitServer;
 import pers.acp.ftp.conf.SFTPConfig;
+import pers.acp.ftp.conf.SFTPListener;
 import pers.acp.ftp.server.SFTPServer;
 import pers.acp.ftp.user.UserFactory;
 
@@ -15,7 +17,7 @@ import java.util.List;
  * Created by zhangbin on 2016/12/21.
  * 启动SFTP服务
  */
-public class InitSFtpServer {
+public class InitSFtpServer extends InitServer {
 
     /**
      * 日志对象
@@ -30,21 +32,19 @@ public class InitSFtpServer {
             sftpServers = doStart(sftpConfig);
         } catch (Exception e) {
             log.error(e.getMessage(), e);
-        } finally {
-            log.info("start sftp servers finished!");
         }
         return sftpServers;
     }
 
-    public static List<IDaemonService> startSFtpServer(SFTPConfig sftpConfig) {
+    public static List<IDaemonService> startSFtpServer(SFTPConfig sftpConfig, List<UserFactory> userFactoryList) {
         log.info("start sftp servers ...");
+        userFactoryList.forEach(InitServer::addUserFactory);
         List<IDaemonService> sftpServers = new ArrayList<>();
         try {
             sftpServers = doStart(sftpConfig);
         } catch (Exception e) {
             log.error(e.getMessage(), e);
-        } finally {
-            log.info("start sftp servers finished!");
+            log.info("start sftp servers exception: " + e.getMessage());
         }
         return sftpServers;
     }
@@ -52,19 +52,18 @@ public class InitSFtpServer {
     private static List<IDaemonService> doStart(SFTPConfig sftpConfig) throws ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InvocationTargetException, InstantiationException {
         List<IDaemonService> sftpServers = new ArrayList<>();
         if (sftpConfig != null) {
-            List<SFTPConfig.Listen> listens = sftpConfig.getListens();
+            List<SFTPListener> listens = sftpConfig.getListens();
             if (listens != null) {
-                for (SFTPConfig.Listen listen : listens) {
+                for (SFTPListener listen : listens) {
                     if (listen.isEnabled()) {
                         String classname = listen.getUserFactoryClass();
                         if (!CommonTools.isNullStr(classname)) {
-                            UserFactory userFactory = (UserFactory) Class.forName(classname).getDeclaredConstructor().newInstance();
+                            UserFactory userFactory = getUserFactory(classname);
                             SFTPServer sftpServer = new SFTPServer(userFactory.generateSFtpUserList(), listen);
                             Thread sub = new Thread(sftpServer);
                             sub.setDaemon(true);
                             sub.start();
                             sftpServers.add(sftpServer);
-                            log.info("start sftp server success [" + listen.getName() + "] port:" + listen.getPort());
                         } else {
                             log.info("start sftp server failed [" + listen.getName() + "] : user factory class is null");
                         }