+++ core/acp-ftp/src/main/java/pers/acp/ftp/server/UserPasswordAuthenticator.java
@@ -1,6 +1,6 @@
 package pers.acp.ftp.server;
 
-import org.apache.sshd.server.PasswordAuthenticator;
+import org.apache.sshd.server.auth.password.PasswordAuthenticator;
 import org.apache.sshd.server.session.ServerSession;
 import pers.acp.core.log.LogFactory;